requirements.txt
PySide2==5.15.0
PyVISA==1.10.1
pyzmq==19.0.1
matplotlib==3.3.1
pytest==6.0.2